refactor PointerObject things into its own .cpp/.hpp file
